Close #457 ✨ don't allow default release to be changed if config file doesn't exist

+++ src/helpers/Releases/setDefaultRelease.js
@@ -4,21 +4,27 @@ const path = require('path');
 // set default release version in godot hub config
 const setDefaultRelease = (godotHubPath, defaultRelease) => {
   const configFilePath = path.join(godotHubPath, '.config', '.config.json');
-  const readConfigFile = JSON.parse(fs.readFileSync(configFilePath));
 
-  // check if there is a default release version in the config file
-  if (readConfigFile.hasOwnProperty('defaultGodotVersion')) {
-    const defaultGodotVersion = readConfigFile.defaultGodotVersion;
+  // check if config file exists before attempting to read and modify it
+  if (fs.existsSync(configFilePath)) {
+    const readConfigFile = JSON.parse(fs.readFileSync(configFilePath));
 
-    // modify default godot version only if its different
-    if (String(defaultGodotVersion) !== String(defaultRelease)) {
+    // check if there is a default release version in the config file
+    if (readConfigFile.hasOwnProperty('defaultGodotVersion')) {
+      const defaultGodotVersion = readConfigFile.defaultGodotVersion;
+
+      // modify default godot version only if its different
+      if (String(defaultGodotVersion) !== String(defaultRelease)) {
+        readConfigFile.defaultGodotVersion = defaultRelease;
+
+        fs.writeFileSync(configFilePath, JSON.stringify(readConfigFile, null, 2));
+      }
+    } else {
       readConfigFile.defaultGodotVersion = defaultRelease;
-
       fs.writeFileSync(configFilePath, JSON.stringify(readConfigFile, null, 2));
     }
   } else {
-    readConfigFile.defaultGodotVersion = defaultRelease;
-    fs.writeFileSync(configFilePath, JSON.stringify(readConfigFile, null, 2));
+    return false;
   }
 };
